Refactor SQL queries: enforce parameterized bindings and consistent style

+++ models/DataObject/ClassDefinition/Data/AdvancedManyToManyObjectRelation.php
@@ -16,6 +16,7 @@ declare(strict_types=1);
 
 namespace OpenDxp\Model\DataObject\ClassDefinition\Data;
 
+use Doctrine\DBAL\ArrayParameterType;
 use Exception;
 use OpenDxp;
 use OpenDxp\Bundle\AdminBundle\Service\GridData;
@@ -122,7 +123,9 @@ class AdvancedManyToManyObjectRelation extends ManyToManyObjectRelation implemen
             }
 
             $existingTargets = $db->fetchFirstColumn(
-                'SELECT id FROM objects WHERE id IN ('.implode(',', $targets).')'
+                'SELECT id FROM objects WHERE id IN (?)',
+                [$targets],
+                [ArrayParameterType::INTEGER]
             );
 
             foreach ($data as $key => $relation) {
@@ -456,26 +459,41 @@ class AdvancedManyToManyObjectRelation extends ManyToManyObjectRelation implemen
                 $ownerName = '/' . $context['containerType'] . '~' . $containerName . '/%';
             }
 
-            $sql = Db\Helper::quoteInto($db, 'id = ?', $objectId) . " AND ownertype = 'localizedfield' AND "
-                . Db\Helper::quoteInto($db, 'ownername LIKE ?', $ownerName)
-                . ' AND ' . Db\Helper::quoteInto($db, 'fieldname = ?', $this->getName())
-                . ' AND ' . Db\Helper::quoteInto($db, 'position = ?', $position);
+            $qb = $db->createQueryBuilder()
+                ->delete($table)
+                ->where('id = :id')
+                ->andWhere('ownertype = "localizedfield"')
+                ->andWhere('ownername LIKE :ownerName')
+                ->andWhere('fieldname = :fieldname')
+                ->andWhere('position = :position')
+                ->setParameter('id', $objectId)
+                ->setParameter('ownerName', $ownerName)
+                ->setParameter('fieldname', $this->getName())
+                ->setParameter('position', $position);
         } else {
-            $sql = Db\Helper::quoteInto($db, 'id = ?', $objectId) . ' AND ' . Db\Helper::quoteInto($db, 'fieldname = ?', $this->getName())
-                . ' AND ' . Db\Helper::quoteInto($db, 'position = ?', $position);
+            $qb = $db->createQueryBuilder()
+                ->delete($table)
+                ->where('id = :id')
+                ->andWhere('fieldname = :fieldname')
+                ->andWhere('position = :position')
+                ->setParameter('id', $objectId)
+                ->setParameter('fieldname', $this->getName())
+                ->setParameter('position', $position);
 
             if ($context) {
                 if (!empty($context['fieldname'])) {
-                    $sql .= ' AND '.Db\Helper::quoteInto($db, 'ownername = ?', $context['fieldname']);
+                    $qb->andWhere('ownername = :ownername')
+                       ->setParameter('ownername', $context['fieldname']);
                 }
 
                 if (!DataObject::isDirtyDetectionDisabled() && $context['containerType']) {
-                    $sql .= ' AND '.Db\Helper::quoteInto($db, 'ownertype = ?', $context['containerType']);
+                    $qb->andWhere('ownertype = :ownertype')
+                       ->setParameter('ownertype', $context['containerType']);
                 }
             }
         }
 
-        $db->executeStatement('DELETE FROM ' . $table . ' WHERE ' . $sql);
+        $qb->executeStatement();
 
         if (!empty($objectsMetadata)) {
             if ($object instanceof DataObject\Localizedfield || $object instanceof DataObject\Objectbrick\Data\AbstractData
@@ -537,10 +555,8 @@ class AdvancedManyToManyObjectRelation extends ManyToManyObjectRelation implemen
             $containerName = $context['fieldname'] ?? null;
             $index = $context['index'];
             $db->executeStatement(
-                'DELETE FROM object_metadata_' . $object->getClassId()
-                . ' WHERE ' . Db\Helper::quoteInto($db, 'id = ?', $object->getId()) . " AND ownertype = 'localizedfield' AND "
-                . Db\Helper::quoteInto($db, 'ownername LIKE ?', '/' . $context['containerType'] . '~' . $containerName . '/' . "$index . /%")
-                . ' AND ' . Db\Helper::quoteInto($db, 'fieldname = ?', $this->getName())
+                sprintf('DELETE FROM object_metadata_%s WHERE id = ? AND ownertype = "localizedfield" AND ownername LIKE ? AND fieldname = ?', $object->getClassId()),
+                [$object->getId(), '/' . $context['containerType'] . '~' . $containerName . '/' . $index . '/%', $this->getName()]
             );
         } else {
             $deleteConditions = [
